Split the Authentication functionality into a Concern
We need the same functionality to serve files with
Active Storage
